requirements.txt
streamlit>=1.28.0
openai>=1.0.0
fpdf2>=2.7.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0